Altera estrutura de pastas para utilizar AngularJS.
