+++ MatheusEduardo/public/webdev-2.js
@@ -1,0 +1,18 @@
+'use strict';
+
+angular.module('webdev-2', [])
+  
+  .controller('mainController', ($scope) => {
+
+    $scope.adicionarUsuario = () => {
+
+    }
+
+    $scope.excluirUsuario = () => {
+
+    }
+
+    $scope.editarUsuario = () => {
+
+    }
+  });
